Release/uat/v3.0.0 2026 05 11

+++ Modules/Appraisal/Appraisal/Application/Features/Appraisals/GetMachineryProperty/GetMachineryPropertyResponse.cs
@@ -45,7 +45,7 @@ public record GetMachineryPropertyResponse(
     string? Location,
     string? ConditionUse,
     string? MachineCondition,
-    int? MachineAge,
+    decimal? MachineAge,
     string? MachineEfficiency,
     string? MachineTechnology,
     string? UsagePurpose,
